return types, 64-bit changes: ioctl cmd -> u_long, casts, etc.

+++ sys/kern/tty_tty.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)tty_tty.c	8.2 (Berkeley) 9/23/93
+ *	@(#)tty_tty.c	8.3 (Berkeley) 1/9/95
  */
 
 /*
@@ -88,7 +88,7 @@ cttywrite(dev, uio, flag)
 /*ARGSUSED*/
 cttyioctl(dev, cmd, addr, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t addr;
 	int flag;
 	struct proc *p;
